Incr 6 - Рефакторинг
