+++ internal/agent/collector/collector.go
@@ -1,0 +1,7 @@
+package collector
+
+import "github.com/m1khal3v/gometheus/internal/metric"
+
+type Collector interface {
+	Collect() []metric.Metric
+}
