--- internal/route/route_container.go
@@ -1,15 +1,0 @@
-package route
-
-import (
-	"github.com/m1khal3v/gometheus/internal/storage"
-)
-
-type Container struct {
-	Storage storage.Storage
-}
-
-func NewRouteContainer(storage storage.Storage) *Container {
-	return &Container{
-		Storage: storage,
-	}
-}
