+++ internal/server/api/get_metric.go
@@ -1,0 +1,20 @@
+package api
+
+import (
+	"net/http"
+)
+
+func (container Container) GetMetric(writer http.ResponseWriter, request *http.Request) {
+	metricType := request.PathValue("type")
+	metricName := request.PathValue("name")
+
+	metric := container.manager.Get(metricName)
+	if metric == nil || metric.GetType() != metricType {
+		writer.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	writer.Header().Set("Content-Type", "text/plain")
+	writer.WriteHeader(http.StatusOK)
+	_, _ = writer.Write([]byte(metric.GetStringValue()))
+}
