+++ internal/server/server.go
@@ -2,8 +2,8 @@ package server
 
 import (
 	"github.com/m1khal3v/gometheus/internal/logger"
-	"github.com/m1khal3v/gometheus/internal/router"
-	"github.com/m1khal3v/gometheus/internal/storage/memory"
+	"github.com/m1khal3v/gometheus/internal/server/router"
+	"github.com/m1khal3v/gometheus/internal/server/storage/memory"
 	"net/http"
 )
