+++ internal/server/storage/storage.go
@@ -1,0 +1,11 @@
+package storage
+
+import (
+	"github.com/m1khal3v/gometheus/internal/metric"
+)
+
+type Storage interface {
+	Save(metric metric.Metric)
+	Get(name string) metric.Metric
+	GetAll() map[string]metric.Metric
+}
